Applied patch from Stefan Huehner (OFBIZ-511), to remove some unnecessary casts where the source-type and the casted-to type are identical.
git-svn-id: https://svn.apache.org/repos/asf/incubator/ofbiz/trunk@481546 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/order/src/org/ofbiz/order/finaccount/FinAccountHelper.java
@@ -109,7 +109,7 @@ public class FinAccountHelper {
          while (!foundUniqueNewCode) {
             newAccountCode = new StringBuffer(codeLength);
             for (int i = 0; i < codeLength; i++) {
-                newAccountCode.append(char_pool[(int) r.nextInt(char_pool.length)]);
+                newAccountCode.append(char_pool[r.nextInt(char_pool.length)]);
             }
 
      	    List existingAccountsWithCode = delegator.findByAnd("FinAccount", UtilMisc.toMap("finAccountCode", newAccountCode.toString()));
